+++ routes/index.js
@@ -5,6 +5,6 @@ export const indexRouter = express.Router()
 /* GET home page. */
 indexRouter.get('/', function (req, res, next) {
   res.render('index', {
-    title: 'Tron',
+    title: 'Bitcycles',
   })
 })